backend/catalog/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
pydantic>=2.6
redis>=5
alembic
psycopg2-binary

httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
